--- monitoring/uss_qualifier/scenarios/astm/utm/data_exchange_validation/test_steps/query_mock_uss_interactions.md
@@ -1,8 +1,0 @@
-# Query mock_uss interactions test step fragment
-
-This step obtains interactions of interest from mock_uss.
-
-## 🛑 Mock USS interactions logs retrievable check
-
-If the query to mock_uss fails or uss_qualifier is otherwise unable to retrieve the interactions of interest, the mock_uss provider does not meet
-**[interuss.mock_uss.hosted_instance.ExposeInterface](../../../../../requirements/interuss/mock_uss/hosted_instance.md)**.
